feat(plugins): implement Phase 2.7 Plugin System
- Add plugin architecture with bundle format and manifest parsing
- Implement PluginLoader for discovering plugins in ~/Library/Application Support
- Add PluginValidator for code signature verification using Security.framework
- Create sandboxed runtime with PluginHost, PluginContextImpl, PluginStorageImpl
- Implement Plugin API: clipboard access, content transform, UI extensions
- Build 5 official plugins: JSONBeautifier, MarkdownFormatter, URLShortener, GitHubGist, Notion
- Create PasteShelfPluginKit Swift package for third-party developers
- Add plugin settings UI (PluginSettingsView)
- Add comprehensive documentation and examples
- Fix test compilation issues (actor isolation, API mismatches, imports)

+++ PasteShelfTests/HotkeyManagerTests.swift
@@ -6,6 +6,7 @@
 //
 
 import Carbon.HIToolbox
+import Foundation
 import Testing
 @testable import PasteShelf
 
@@ -52,13 +53,13 @@ struct HotkeyManagerTests {
     @Test("Display string for common keys")
     func displayStringForCommonKeys() {
         let testCases: [(Int32, String)] = [
-            (kVK_ANSI_A, "A"),
-            (kVK_ANSI_V, "V"),
-            (kVK_ANSI_C, "C"),
-            (kVK_Space, "␣"),
-            (kVK_Return, "↩"),
-            (kVK_Tab, "⇥"),
-            (kVK_Escape, "⎋"),
+            (Int32(kVK_ANSI_A), "A"),
+            (Int32(kVK_ANSI_V), "V"),
+            (Int32(kVK_ANSI_C), "C"),
+            (Int32(kVK_Space), "␣"),
+            (Int32(kVK_Return), "↩"),
+            (Int32(kVK_Tab), "⇥"),
+            (Int32(kVK_Escape), "⎋"),
         ]
 
         for (keyCode, expectedKey) in testCases {
